[Refactor] Middleware separation: split idempotency/rate-limit/request-size into composable units
