+++ api/main.py
@@ -16,14 +16,7 @@ from fastapi import status
 import structlog
 
 from api.config import get_settings
-from api.middleware import (
-    rate_limit_middleware,
-    add_correlation_id_middleware,
-    error_handling_middleware,
-    logging_middleware,
-    request_size_limit_middleware,
-    idempotency_middleware,
-)
+from api.middlewares import register_middlewares
 from api.csrf import CSRFProtectionMiddleware
 from api.limiter import cleanup_limiter
 from observability.integration import initialize_observability_for_environment
@@ -155,16 +148,9 @@ def create_app() -> FastAPI:
     
     # Initialize validation config from settings
     ValidationConfig.from_settings(settings)
-    
-    # Add middleware
-    app.middleware("http")(request_size_limit_middleware)
-    app.middleware("http")(idempotency_middleware)
-    app.middleware("http")(add_correlation_id_middleware)
-    app.middleware("http")(logging_middleware)
-    app.middleware("http")(error_handling_middleware)
 
-    if settings.RATE_LIMIT_ENABLED:
-        app.middleware("http")(rate_limit_middleware)
+    # Add middleware through a single registration entry to preserve order.
+    register_middlewares(app)
 
     if _GRAPHQL_ROUTER is not None:
         app.include_router(
