+++ db/models/auth.py
@@ -1,5 +1,7 @@
 import datetime as dt
-from sqlalchemy import Column, Integer, String, DateTime, Boolean, ForeignKey
+import enum
+
+from sqlalchemy import Column, Integer, String, DateTime, Boolean, ForeignKey, Enum as SQLEnum
 from sqlalchemy.orm import relationship
 from db.base import Base
